Remove Entity subclasses
 - Any Entity-specific functionality should be added via components

 - Combine UnitRenderer and BuildingRenderer into EntityRenderer

 - Replace Unit / Building classes with namespaces and utility methods

 - Add SpriteComponent instead of tracking the Renderables in the renderer

 - Move unit facings into FacingComponent

+++ Open-Rival/include/Camera.h
@@ -22,8 +22,8 @@ namespace Rival {
 
     public:
         // Size of a tile, in camera co-ordinates
-        static const float tileWidth;
-        static const float tileHeight;
+        static const int tileWidth;
+        static const int tileHeight;
 
         // Zoom behaviour
         static const float zoomInterval;
